+++ alert/types.go
@@ -17,6 +17,18 @@ type Event struct {
 	previousState EventState
 }
 
+func (e Event) AlertData() Data {
+	return Data{
+		ID:       e.State.ID,
+		Message:  e.State.Message,
+		Details:  e.State.Details,
+		Time:     e.State.Time,
+		Duration: e.State.Duration,
+		Level:    e.State.Level,
+		Data:     e.Data.Result,
+	}
+}
+
 func (e Event) PreviousState() EventState {
 	return e.previousState
 }
@@ -153,4 +165,16 @@ func ParseLevel(s string) (l Level, err error) {
 type TopicState struct {
 	Level     Level
 	Collected int64
+}
+
+// Data is a structure that contains relevant data about an alert event.
+// The structure is intended to be JSON encoded, providing a consistent data format.
+type Data struct {
+	ID       string        `json:"id"`
+	Message  string        `json:"message"`
+	Details  string        `json:"details"`
+	Time     time.Time     `json:"time"`
+	Duration time.Duration `json:"duration"`
+	Level    Level         `json:"level"`
+	Data     models.Result `json:"data"`
 }