+++ pipeline/http_post.go
@@ -1,6 +1,14 @@
 package pipeline
 
+import (
+	"errors"
+	"fmt"
+	"strings"
+)
+
 // An HTTPPostNode will take the incoming data stream and POST it to an HTTP endpoint.
+// That endpoint may be specified as a positional argument, or as an endpoint property
+// method on httpPost. Multiple endpoint property methods may be specified.
 //
 // Example:
 //    stream
@@ -11,16 +19,99 @@ package pipeline
 //        //Post the top 10 results over the last 10s updated every 5s.
 //        |httpPost('http://example.com/api/top10')
 //
+// Example:
+//    stream
+//        |window()
+//            .period(10s)
+//            .every(5s)
+//        |top('value', 10)
+//        //Post the top 10 results over the last 10s updated every 5s.
+//        |httpPost()
+//            .endpoint('example')
+//
 type HTTPPostNode struct {
 	chainnode
 
 	// tick:ignore
-	Url string
+	HTTPPostEndpoints []*HTTPPostEndpoint `tick:"Endpoint"`
+
+	// Headers
+	Headers map[string]string `tick:"Header"`
+
+	// tick:ignore
+	URLs []string
 }
 
-func newHTTPPostNode(wants EdgeType, url string) *HTTPPostNode {
+func newHTTPPostNode(wants EdgeType, urls ...string) *HTTPPostNode {
 	return &HTTPPostNode{
 		chainnode: newBasicChainNode("http_post", wants, wants),
-		Url:       url,
+		URLs:      urls,
 	}
+}
+
+// tick:ignore
+func (p *HTTPPostNode) validate() error {
+	if len(p.URLs) >= 2 {
+		return fmt.Errorf("httpPost expects 0 or 1 arguments, got %v", len(p.URLs))
+	}
+
+	if len(p.HTTPPostEndpoints) > 1 {
+		return fmt.Errorf("httpPost expects 0 or 1 endpoints, got %v", len(p.HTTPPostEndpoints))
+	}
+
+	if len(p.URLs) == 0 && len(p.HTTPPostEndpoints) == 0 {
+		return errors.New("must provide url or endpoint")
+	}
+
+	if len(p.URLs) > 0 && len(p.HTTPPostEndpoints) > 0 {
+		return errors.New("only one endpoint and url may be specified")
+	}
+
+	for k := range p.Headers {
+		if strings.ToUpper(k) == "AUTHENTICATE" {
+			return errors.New("cannot set 'authenticate' header")
+		}
+	}
+
+	return nil
+}
+
+// Example:
+//    stream
+//         |httpPost()
+//            .endpoint('example')
+//
+// tick:property
+func (p *HTTPPostNode) Endpoint(endpoint string) *HTTPPostEndpoint {
+	post := &HTTPPostEndpoint{
+		HTTPPostNode: p,
+		Endpoint:     endpoint,
+	}
+	p.HTTPPostEndpoints = append(p.HTTPPostEndpoints, post)
+
+	return post
+}
+
+// Example:
+//    stream
+//         |httpPost()
+//            .endpoint('example')
+//              .header('my', 'header')
+//
+// tick:property
+func (p *HTTPPostNode) Header(k, v string) *HTTPPostNode {
+	if p.Headers == nil {
+		p.Headers = map[string]string{}
+	}
+	p.Headers[k] = v
+
+	return p
+}
+
+// tick:embedded:HTTPPostNode.Endpoint
+type HTTPPostEndpoint struct {
+	*HTTPPostNode
+
+	// Name of the endpoint to be used, as is defined in the configuration file
+	Endpoint string
 }